requirements.txt
Flask==3.0.3
flask_wtf==1.2.1
numpy==2.1.2
opencv_python==4.10.0.84
Pillow==11.0.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
WTForms==3.1.2
